fix: pin IBLBMFluidNode to float32 — dtype-stable under jax_enable_x64
The LBM solver is float32 by design (init_equilibrium forces float32
with an explicit comment about the x64 leak). But IBLBMFluidNode.update()
read body_angular_velocity straight from boundary_inputs: when a graph
runs under jax_enable_x64, that input is a float64 array, so omega_z —
and through it wall_vel and the streamed distribution f — were silently
promoted to float64.

This made the node's output depend on a global JAX config flag.
test_update_matches_sweep_logic caught it: node f was float64, the
test's manual float32 reference diverged by 2.5e-5 (> atol 1e-6).

Fix: cast body_angular_velocity to float32 on entry to update(), and
pin initial_state()'s body_angle / drag_force / drag_torque to float32.
The node now produces identical float32 results regardless of the
global x64 flag. Verified passing in both default and JAX_ENABLE_X64=1.

+++ src/mime/nodes/environment/lbm/fluid_node.py
@@ -248,9 +248,9 @@ class IBLBMFluidNode(MimeNode):
         # Neither belongs in the state pytree.
         return {
             "f": init_equilibrium(nx, ny, nz),
-            "body_angle": jnp.array(0.0),
-            "drag_force": jnp.zeros(3),
-            "drag_torque": jnp.zeros(3),
+            "body_angle": jnp.array(0.0, dtype=jnp.float32),
+            "drag_force": jnp.zeros(3, dtype=jnp.float32),
+            "drag_torque": jnp.zeros(3, dtype=jnp.float32),
         }
 
     def boundary_input_spec(self) -> dict[str, BoundaryInputSpec]:
@@ -278,9 +278,13 @@ class IBLBMFluidNode(MimeNode):
         nz = self.params["nz"]
         center = self._center
 
+        # The LBM solver is float32 by design (see init_equilibrium's
+        # explicit dtype).  Cast the boundary input so a graph running
+        # under jax_enable_x64 cannot silently promote wall_vel — and
+        # therefore the streamed f — to float64.
         omega_vec = boundary_inputs.get(
             "body_angular_velocity", jnp.zeros(3),
-        )
+        ).astype(jnp.float32)
         omega_z = omega_vec[2]
 
         # 1. Update angle (dt_lbm = 1 in lattice units)
